% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_rules.R
\name{antiplurality_all_winners}
\alias{antiplurality_all_winners}
\title{Return all anti-plurality winners}
\usage{
antiplurality_all_winners(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Integer vector of 1-based alternative indices.
}
\description{
Return all anti-plurality winners
}
