% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_rules.R
\name{antiplurality_scores}
\alias{antiplurality_scores}
\title{Compute anti-plurality scores}
\usage{
antiplurality_scores(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Integer vector of length n_alts.
}
\description{
\code{scores[j]} = number of voters for whom alternative \code{j} is NOT
ranked last. Higher is better.
}
\examples{
\dontrun{
prof <- profile_build_spatial(c(0, 0, 2, 0, -2, 0), c(-1, -1, 1, -1, 1, 1))
antiplurality_scores(prof)
}
}
