% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_rules.R
\name{approval_all_winners_topk}
\alias{approval_all_winners_topk}
\title{Return all winners under the top-k approval rule}
\usage{
approval_all_winners_topk(profile, k)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}

\item{k}{Integer in \code{[1, n_alts]}.}
}
\value{
Integer vector of 1-based alternative indices.
}
\description{
Return all winners under the top-k approval rule
}
