% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_rules.R
\name{borda_all_winners}
\alias{borda_all_winners}
\title{Return all Borda winners}
\usage{
borda_all_winners(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Integer vector of 1-based alternative indices.
}
\description{
Return all Borda winners
}
