% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{.PALETTE_HEX}
\alias{.PALETTE_HEX}
\title{Approved qualitative palettes.
Source: ColorBrewer 2.0 (Cynthia Brewer) and Okabe & Ito (2008).}
\format{
An object of class \code{list} of length 4.
}
\usage{
.PALETTE_HEX
}
\description{
Approved qualitative palettes.
Source: ColorBrewer 2.0 (Cynthia Brewer) and Okabe & Ito (2008).
}
\keyword{datasets}
