% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{is_pareto_efficient}
\alias{is_pareto_efficient}
\title{Test whether a single alternative is Pareto-efficient}
\usage{
is_pareto_efficient(profile, alt)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}

\item{alt}{Integer (1-based). The alternative to test.}
}
\value{
Logical scalar.
}
\description{
Test whether a single alternative is Pareto-efficient
}
