% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{list_scenarios}
\alias{list_scenarios}
\title{List available built-in scenarios}
\usage{
list_scenarios()
}
\value{
Character vector of scenario identifiers.
}
\description{
Returns the names that can be passed to \code{load_scenario()}.
}
\examples{
list_scenarios()

}
