% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voting_rules.R
\name{plurality_scores}
\alias{plurality_scores}
\title{Compute plurality scores (first-place vote counts)}
\usage{
plurality_scores(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Integer vector of length n_alts. Names are \code{"alt1", "alt2", ...}.
}
\description{
Compute plurality scores (first-place vote counts)
}
\examples{
\dontrun{
alts   <- c(0, 0, 2, 0, -2, 0)
voters <- c(-1, -1, 1, -1, 1, 1)
prof   <- profile_build_spatial(alts, voters)
plurality_scores(prof)
}
}
