% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scs_version}
\alias{scs_version}
\title{Query the compiled C API version}
\usage{
scs_version()
}
\value{
Named list \code{list(major, minor, patch)}.
}
\description{
Query the compiled C API version
}
\examples{
\dontrun{
# Requires libscs_api to be built. See the package README.
scs_version()
}
}
