% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{stream_manager}
\alias{stream_manager}
\title{Create a StreamManager (convenience factory)}
\usage{
stream_manager(master_seed, stream_names = NULL)
}
\arguments{
\item{master_seed}{Integer or double. Master seed.}

\item{stream_names}{Optional character vector of stream names to register
immediately.}
}
\value{
A \code{\link{StreamManager}} object.
}
\description{
Equivalent to \code{StreamManager$new(master_seed, stream_names)}.
}
