% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{condorcet_winner_profile}
\alias{condorcet_winner_profile}
\title{Return the Condorcet winner from a profile}
\usage{
condorcet_winner_profile(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Integer (1-based), or \code{NA_integer_} if no Condorcet winner.
}
\description{
Return the Condorcet winner from a profile
}
