% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{has_condorcet_winner_profile}
\alias{has_condorcet_winner_profile}
\title{Test whether the profile has a Condorcet winner}
\usage{
has_condorcet_winner_profile(profile)
}
\arguments{
\item{profile}{A \code{\link{Profile}} object.}
}
\value{
Logical scalar.
}
\description{
A Condorcet winner beats every other alternative by strict majority (more
than half of voters).
}
